+++ examples/node/singleLocale.js
@@ -1,7 +1,7 @@
-var pure = require('../locale/en');
+var pure = require('../../locale/en');
 
 console.log(pure.name.findName());
 
-var pure = require('../locale/uk');
+var pure = require('../../locale/uk');
 
 console.log(pure.name.findName());